Transfer receiving localStorage date to App.
